On-demand QPs setup
