+++ examples/cc_main.c
@@ -71,7 +71,8 @@ struct cc_conf {
         int                     use_mq;
         int                     oob_barrier;
 	int                     size;
-	struct cc_alg_info     *algorithm;
+        struct cc_alg_info     *algorithm;
+        int on_demand_conn;
 };
 
 #pragma pack( push, 1 )
@@ -175,6 +176,9 @@ static inline void __timer(struct cc_timer *t)
 			(ru.ru_utime.tv_usec + ru.ru_stime.tv_usec);
 }
 
+
+
+
 static inline int __log2_cc(long val)
 {
 	int count = 0;
@@ -231,23 +235,7 @@ static int __repost(struct cc_context *ctx, struct ibv_qp *qp, int count, int pe
     return posted;
 }
 
-#if defined(USE_MPI)
-static int __exchange_info_by_mpi(struct cc_context *ctx,
-				  struct cc_proc_info *my_info,
-				  struct cc_proc_info *peer_info,
-				  int peer_proc)
-{
-	MPI_Status *status = NULL;
-
-	MPI_Sendrecv(my_info, sizeof(*my_info), MPI_CHAR,
-		     peer_proc, 0,
-		     peer_info, sizeof(*peer_info), MPI_CHAR,
-		     peer_proc, 0,
-		     MPI_COMM_WORLD, status);
-
-	return 0;
-}
-#endif
+#include "cc_utils.h"
 
 static int __init_ctx( struct cc_context *ctx )
 {
@@ -401,188 +389,18 @@ static int __init_ctx( struct cc_context *ctx )
 	 * =======================================
 	 */
 	for (i = 0; i < ctx->conf.num_proc; i++) {
-
-                // if (i == ctx->conf.my_proc)
-                        // continue;
-
-		/*
+            if (ctx->conf.on_demand_conn) {
+                ctx->proc_array[i].rcq = NULL;
+            }else{
+                /*
 		 * 4.1 Create QPs for Peers
 		 * =======================================
 		 */
 		log_trace("create QPs for peers ...\n");
-		{
-			struct ibv_exp_cq_attr attr;
-			struct ibv_exp_qp_init_attr init_attr;
-
-                        ctx->proc_array[i].rcq = ibv_create_cq(ctx->ib_ctx, ctx->conf.cq_rx_depth, NULL, NULL, 0);
-                        ctx->proc_array[i].scq = ctx->proc_array[i].rcq;//ctx->proc_array[ctx->conf.my_proc].scq;
-
-			if (!ctx->proc_array[i].rcq) {
-				log_fatal("ibv_create_cq failed\n");
-			}
-
-			log_trace("create QPs for peers ...scq=%p  rcq=%p\n", ctx->proc_array[i].scq, ctx->proc_array[i].rcq);
-
-
-                        attr.comp_mask            = IBV_EXP_CQ_ATTR_CQ_CAP_FLAGS;
-                        attr.moderation.cq_count  = 0;
-			attr.moderation.cq_period = 0;
-			attr.cq_cap_flags         = IBV_EXP_CQ_IGNORE_OVERRUN;
-
-			rc = ibv_exp_modify_cq(ctx->proc_array[i].rcq, &attr, IBV_EXP_CQ_CAP_FLAGS);
-			if (rc) {
-				log_fatal("ibv_modify_cq failed\n");
-			}
-
-			memset(&init_attr, 0, sizeof(init_attr));
-
-			init_attr.qp_context = NULL;
-			init_attr.send_cq = ctx->proc_array[i].scq;
-			init_attr.recv_cq = ctx->proc_array[i].rcq;
-			init_attr.srq = NULL;
-			init_attr.cap.max_send_wr  = ctx->conf.qp_tx_depth;
-			init_attr.cap.max_recv_wr  = ctx->conf.qp_rx_depth;
-			init_attr.cap.max_send_sge = 16;
-			init_attr.cap.max_recv_sge = 16;
-                        init_attr.cap.max_inline_data = 256;
-			init_attr.qp_type = IBV_QPT_RC;
-                        init_attr.sq_sig_all = 0;
-			init_attr.pd = ctx->pd;
-
-			{
-                                init_attr.comp_mask |= IBV_EXP_QP_INIT_ATTR_CREATE_FLAGS | IBV_EXP_QP_INIT_ATTR_PD;
-                                if (ctx->conf.use_mq) {
-                                    init_attr.exp_create_flags =
-                                        IBV_EXP_QP_CREATE_CROSS_CHANNEL      |
-                                        IBV_EXP_QP_CREATE_MANAGED_SEND       |
-                                        IBV_EXP_QP_CREATE_IGNORE_SQ_OVERFLOW |
-                                        IBV_EXP_QP_CREATE_IGNORE_RQ_OVERFLOW;
-                                } else {
-                                    init_attr.exp_create_flags =
-                                        IBV_EXP_QP_CREATE_CROSS_CHANNEL      |
-                                        IBV_EXP_QP_CREATE_IGNORE_SQ_OVERFLOW |
-                                        IBV_EXP_QP_CREATE_IGNORE_RQ_OVERFLOW;
-;
-                                }
-				ctx->proc_array[i].qp = ibv_exp_create_qp(ctx->ib_ctx, &init_attr);
-			}
-			if (!ctx->proc_array[i].qp)
-				log_fatal("ibv_create_qp_ex failed\n");
-
-		}
-	}
-		for (i = 0; i < ctx->conf.num_proc; i++) {
-			struct cc_proc_info local_info;
-			struct cc_proc_info remote_info;
-
-                        // if (i == ctx->conf.my_proc)
-                                // continue;
-		/*
-		 * 4.2 Exchange information with Peers
-		 * =======================================
-		 */
-		log_trace("exchange info with peers ...\n");
-		{
-			local_info.lid = htons(__get_local_lid(ctx->ib_ctx, ctx->ib_port));
-			local_info.qpn = htonl(ctx->proc_array[i].qp->qp_num);
-			local_info.psn = htonl(lrand48() & 0xffffff);
-			local_info.vaddr = htonll((uintptr_t)ctx->buf + ctx->conf.size * (i + 1));
-			local_info.rkey = htonl(ctx->mr->rkey);
-
-#if defined(USE_MPI)
-			rc = __exchange_info_by_mpi(ctx, &local_info, &remote_info, i);
-#endif
-			ctx->proc_array[i].info.lid = ntohs(remote_info.lid);
-			ctx->proc_array[i].info.qpn = ntohl(remote_info.qpn);
-			ctx->proc_array[i].info.psn = ntohl(remote_info.psn);
-			ctx->proc_array[i].info.vaddr = ntohll(remote_info.vaddr);
-			ctx->proc_array[i].info.rkey = ntohl(remote_info.rkey);
-
-			log_trace("local     info: lid = %d qpn = %d psn = %d vaddr = 0x%lx rkey = %d\n",
-					ntohs(local_info.lid),
-					ntohl(local_info.qpn),
-					ntohl(local_info.psn),
-					ntohll(local_info.vaddr),
-					ntohl(local_info.rkey));
-			log_trace("peer(#%d) info: lid = %d qpn = %d psn = %d vaddr = 0x%lx rkey = %d\n",
-					i,
-					ctx->proc_array[i].info.lid,
-					ctx->proc_array[i].info.qpn,
-					ctx->proc_array[i].info.psn,
-					ctx->proc_array[i].info.vaddr,
-					ctx->proc_array[i].info.rkey);
-		}
-
-		/*
-		 * 4.3 Complete QPs for Peers
-		 * =======================================
-		 */
-		log_trace("connect to peers ...\n");
-		{
-			struct ibv_qp_attr attr;
-
-                        if (__repost(ctx, ctx->proc_array[i].qp, ctx->conf.qp_rx_depth,i) != ctx->conf.qp_rx_depth)
-				log_fatal("__post_read failed\n");
-
-			memset(&attr, 0, sizeof(attr));
-
-			attr.qp_state        = IBV_QPS_INIT;
-			attr.pkey_index      = 0;
-                        attr.port_num        = ctx->ib_port;
-                        attr.qp_access_flags = IBV_ACCESS_REMOTE_READ | IBV_ACCESS_REMOTE_WRITE | IBV_ACCESS_LOCAL_WRITE;
-
-			rc = ibv_modify_qp(ctx->proc_array[i].qp, &attr,
-					IBV_QP_STATE              |
-					IBV_QP_PKEY_INDEX         |
-					IBV_QP_PORT               |
-					IBV_QP_ACCESS_FLAGS);
-			if (rc)
-				log_fatal("ibv_modify_qp failed\n");
-
-			memset(&attr, 0, sizeof(attr));
-
-			attr.qp_state              = IBV_QPS_RTR;
-			attr.path_mtu              = IBV_MTU_1024;
-			attr.dest_qp_num	   = ctx->proc_array[i].info.qpn;
-			attr.rq_psn                = ctx->proc_array[i].info.psn;
-			attr.max_dest_rd_atomic    = 4;
-			attr.min_rnr_timer         = 12;
-			attr.ah_attr.is_global     = 0;
-			attr.ah_attr.dlid          = ctx->proc_array[i].info.lid;
-			attr.ah_attr.sl            = 0;
-			attr.ah_attr.src_path_bits = 0;
-			attr.ah_attr.port_num      = ctx->ib_port;
-
-			rc = ibv_modify_qp(ctx->proc_array[i].qp, &attr,
-					IBV_QP_STATE              |
-					IBV_QP_AV                 |
-					IBV_QP_PATH_MTU           |
-					IBV_QP_DEST_QPN           |
-					IBV_QP_RQ_PSN             |
-					IBV_QP_MAX_DEST_RD_ATOMIC |
-					IBV_QP_MIN_RNR_TIMER);
-			if (rc)
-				log_fatal("ibv_modify_qp failed\n");
-
-                        memset(&attr, 0, sizeof(attr));
-
-			attr.qp_state      = IBV_QPS_RTS;
-			attr.timeout       = 14;
-			attr.retry_cnt     = 7;
-			attr.rnr_retry     = 7;
-			attr.sq_psn        = ntohl(local_info.psn);
-			attr.max_rd_atomic = 4;
-			rc = ibv_modify_qp(ctx->proc_array[i].qp, &attr,
-					IBV_QP_STATE              |
-					IBV_QP_TIMEOUT            |
-					IBV_QP_RETRY_CNT          |
-					IBV_QP_RNR_RETRY          |
-					IBV_QP_SQ_PSN             |
-					IBV_QP_MAX_QP_RD_ATOMIC);
-			if (rc)
-				log_fatal("ibv_modify_qp failed\n");
-		}
-	}
+                create_qp_to_dest(i, ctx);
+                exchange_qp_info_and_connect(i, ctx);
+            }
+        }
 
 	return rc;
 }
@@ -682,6 +500,7 @@ static int __parse_cmd_line( struct cc_context *ctx, int argc, char * const argv
         ctx->conf.cq_rx_depth = 0x0200;
         ctx->conf.qp_tx_depth = 0x0200;
         ctx->conf.qp_rx_depth = 0x0200;
+        ctx->conf.on_demand_conn = 0;
 	ctx->conf.algorithm = &__barrier_algorithm_recursive_doubling_info;
 
 
@@ -707,11 +526,12 @@ static int __parse_cmd_line( struct cc_context *ctx, int argc, char * const argv
 			{ .name = "check",	.has_arg = 1, .val = 'c' },
 			{ .name = "warmup",	.has_arg = 1, .val = 'w' },
 			{ .name = "size",	.has_arg = 1, .val = 's' },
-			{ .name = "debug",	.has_arg = 1, .val = 'd' },
+                        { .name = "debug",	.has_arg = 1, .val = 'd' },
+                        { .name = "on-demand-conn",	.has_arg = 1, .val = 'o' },
 			{ 0 }
 		};
 
-		c = getopt_long(argc, argv, "?t:ln:r:i:s:d", long_options, NULL);  // added a:l
+                c = getopt_long(argc, argv, "?t:ln:r:i:s:d:o:", long_options, NULL);  // added a:l
 		if (c == -1)
 			break;
 
@@ -756,7 +576,11 @@ static int __parse_cmd_line( struct cc_context *ctx, int argc, char * const argv
 
 		case 'd':
 			__debug_mask = strtol(optarg, NULL, 0);
+                        break;
+                case 'o':
+                        ctx->conf.on_demand_conn = strtol(optarg, NULL, 0);
 			break;
+                        
 
 		default:
 			__usage(argv[0]);
